3. rocnik/semester 6/DBS/requirements.txt
uvicorn
fastapi
psycopg2-binary
tzdata
pydantic-settings
Unidecode
